Move node tests under node submodule
